requirements.txt
django
djongo
pyyaml
opencv-python
torch
torchvision
